[Refactoring] add request Annotation
