+++ src/main/java/space/space_spring/dto/VoiceRoom/GetToken.java
@@ -1,10 +1,13 @@
 package space.space_spring.dto.VoiceRoom;
 
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Null;
 import lombok.Getter;
 
 public class GetToken {
     @Getter
     public static class Request{
+        @NotNull
         private long roomId;
     }
     @Getter
