+++ src/main/java/space/space_spring/dto/VoiceRoom/LiveKitDto/LiveKitSessionResponse.java
@@ -1,6 +1,7 @@
-package space.space_spring.dto.VoiceRoom;
+package space.space_spring.dto.VoiceRoom.LiveKitDto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import space.space_spring.dto.VoiceRoom.LiveKitDto.LiveKitSession;
 
 import java.util.List;
