+++ src/main/java/space/space_spring/service/LiveKitService.java
@@ -9,6 +9,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import retrofit2.Response;
 import space.space_spring.dto.VoiceRoom.*;
+import space.space_spring.dto.VoiceRoom.LiveKitDto.GetTokenRequest;
+import space.space_spring.dto.VoiceRoom.LiveKitDto.LiveKitSession;
+import space.space_spring.dto.VoiceRoom.LiveKitDto.LiveKitSessionResponse;
 
 import java.security.SecureRandom;
 import java.util.List;
